Choose your own adventure story
The story focuses around how people, society, and social media can force people into depression. It shows why it is important to share your problems, with your loved ones or anyone who can help. Also how the support of family, friends, or teachers can help a person battle with depression. If the person does not get appropriate support and help, sometimes that can lead the person to take extreme steps.The story explores the idea of how many times a person is ready to speak about their depression or give various indications of their pain but we as a society have to learn to listen. It also shows the sad truth of how sometimes getting help can also not help.